Initial scenario test harness check-in This repo enables basic SDK scenario testing for the .NET SDK.
The harness is modeled off of runtime's single file test runner that self-discovers tests within its own assembly. An executable is generated for each of the test projects, which should correspond to a local grouping of tests. While the harness may be run as a framework dependent app, in many cases it may be more useful to publish as self-contained using an LKG SDK to avoid using the sdk/runtime being tested to run the harness.

Initially, a subset of the smoke tests are included. Tests can be marked with the trait "Category=Offline" to indicate tests that should pass in offline mode.

This code should be offline source-buildable after xunit is included in source-build-externals.