Support executing the test host with a specific --fx-version and define a default
Allow executing the test host with a specific shared framework version (i.e. 10.0.0-ci).
This is necessary as i.e. in the VMR, the live built shared framework has a `-dev` or `-ci`
suffix which could be determined as "lower" than the toolset SDK's shared framework that the repo
builds with. As an example `10.0.0-ci` and `10.0.0-dev` are considered lower than `10.0.0-preview.1`
by the host (in accordance with semver2).